adds missing performance entry tables, captures performance entry data, makes sure to merge performance metrics with previous performance metrics when updating

+++ api/sqs-data-consumer/database/migrations/20230218010849-create-event-performance-entries.js
@@ -22,7 +22,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       start_time: {
-        type: Sequelize.BIGINT
+        type: Sequelize.DECIMAL
       },
       interaction_id: {
         type: Sequelize.STRING
