+++ api/sqs-data-consumer/database/migrations/20230218011839-create-largest-contentful-paint-performance-entries.js
@@ -22,7 +22,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       start_time: {
-        type: Sequelize.BIGINT
+        type: Sequelize.DECIMAL
       },
       element_identifier: {
         type: Sequelize.STRING
