+++ api/sqs-data-consumer/database/migrations/20230218012656-create-longtask-task-attribution-performance-entries.js
@@ -25,7 +25,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       start_time: {
-        type: Sequelize.BIGINT
+        type: Sequelize.DECIMAL
       },
       container_type: {
         type: Sequelize.STRING
